Repositories were created on Sales Domain layer
