+++ demosales-datasource/room/src/main/java/com/imecatro/demosales/datasource/room/mappers/DataSourceToDomainMappers.kt
@@ -1,8 +1,8 @@
 package com.imecatro.demosales.datasource.room.mappers
 
 import com.imecatro.demosales.datasource.room.entities.ProductRoomEntity
-import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
-import com.imecatro.demosales.domain.products.products.model.ProductUnit
+import com.imecatro.demosales.domain.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.model.ProductUnit
 
 fun List<ProductRoomEntity>.toListDomain(): List<ProductDomainModel> {
     return map { it.toDomain() }
