+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/model/ProductDomainModel.kt
@@ -1,4 +1,4 @@
-package com.imecatro.demosales.domain.products.products.model
+package com.imecatro.demosales.domain.products.model
 
 
 data class ProductDomainModel(
