+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/model/ProductUnit.kt
@@ -1,4 +1,4 @@
-package com.imecatro.demosales.domain.products.products.model
+package com.imecatro.demosales.domain.products.model
 
 
 sealed class ProductUnit(val symbol: String) {
