--- demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/products/usecases/UpdateProductUseCase.kt
@@ -1,11 +1,0 @@
-package com.imecatro.demosales.domain.products.products.usecases
-
-import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
-import com.imecatro.demosales.domain.products.products.repository.ProductsRepository
-
-class UpdateProductUseCase(private val productsRepository: ProductsRepository) {
-    operator fun invoke(product: ProductDomainModel) {
-        productsRepository.updateProduct(product)
-    }
-
-}
