+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/repository/DummyRepository.kt
@@ -1,7 +1,7 @@
-package com.imecatro.demosales.domain.products.products.repository
+package com.imecatro.demosales.domain.products.repository
 
-import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
-import com.imecatro.demosales.domain.products.products.model.ProductUnit
+import com.imecatro.demosales.domain.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.model.ProductUnit
 
 object DummyRepository {
