+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/repository/ProductsRepository.kt
@@ -1,6 +1,6 @@
-package com.imecatro.demosales.domain.products.products.repository
+package com.imecatro.demosales.domain.products.repository
 
-import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.model.ProductDomainModel
 import kotlinx.coroutines.flow.Flow
 
 interface ProductsRepository {
@@ -9,4 +9,5 @@ interface ProductsRepository {
     fun deleteProductById(id :Int?)
     fun updateProduct(product: ProductDomainModel?)
     fun getProductDetailsById(id: Int?): ProductDomainModel?
+    fun getProductsStartingWith(letter: String): List<ProductDomainModel>
 }
