+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/usecases/GetAllProductsUseCase.kt
@@ -1,7 +1,7 @@
-package com.imecatro.demosales.domain.products.products.usecases
+package com.imecatro.demosales.domain.products.usecases
 
-import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
-import com.imecatro.demosales.domain.products.products.repository.ProductsRepository
+import com.imecatro.demosales.domain.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.repository.ProductsRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
