+++ demosales-domain/sales/src/main/java/com/imecatro/demosales/domain/sales/add/AddSaleRepository.kt
@@ -1,0 +1,8 @@
+package com.imecatro.demosales.domain.sales.add
+
+import com.imecatro.demosales.domain.sales.model.SaleModelDomain
+
+
+interface AddSaleRepository {
+    fun createNewSale(sale: SaleModelDomain)
+}
