+++ demosales-domain/sales/src/main/java/com/imecatro/demosales/domain/sales/details/DetailsSaleRepository.kt
@@ -1,0 +1,7 @@
+package com.imecatro.demosales.domain.sales.details
+
+interface DetailsSaleRepository {
+    //details
+    fun getDetailsById(id: Int)
+
+}
