+++ demosales-domain/sales/src/main/java/com/imecatro/demosales/domain/sales/list/AllSalesRepository.kt
@@ -1,0 +1,9 @@
+package com.imecatro.demosales.domain.sales.list
+
+import com.imecatro.demosales.domain.sales.model.SaleModelDomain
+
+interface AllSalesRepository {
+    //list
+    suspend fun getAllSales(): List<SaleModelDomain>
+
+}
